marbleproject/requirements.txt
Django==4.1.4
keras==2.11.0
Pillow==9.2.0
scipy==1.9.1
sqlparse==0.4.3
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
gunicorn==20.1.0